@@ -14,21 +14,30 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "babel-core": "^6.9.1",
-    "babel-loader": "^6.2.4",
-    "babel-preset-react": "^6.5.0",
+    "babel": "^6.5.2",
+    "babel-core": "^6.21.0",
+    "babel-loader": "^6.2.10",
+    "babel-preset-react": "^6.16.0",
+    "babel-preset-stage-3": "^6.17.0",
     "flux": "^2.1.1",
-    "react": "^15.1.0",
+    "react": "^15.4.2",
     "react-addons-css-transition-group": "^15.1.0",
     "react-addons-linked-state-mixin": "^15.1.0",
-    "react-dom": "^15.1.0",
+    "react-dom": "^15.4.2",
     "react-ga": "^2.1.0",
     "react-redux": "^5.0.1",
     "react-router": "^2.4.1",
+    "redux": "^3.6.0",
     "redux-little-router": "^12.1.2",
-    "webpack": "^1.13.1"
+    "redux-logger": "^2.7.4",
+    "redux-thunk": "^2.1.0",
+    "simple-react-modal": "^0.5.1",
+    "webpack": "^1.14.0"
   },
   "devDependencies": {
+    "babel-core": "^6.21.0",
+    "babel-loader": "^6.2.10",
+    "babel-preset-es2015": "^6.18.0",
     "eslint": "^3.13.0",
     "eslint-config-airbnb": "^13.0.0",
     "eslint-plugin-import": "^2.2.0",
