@@ -1,37 +1,30 @@
-var path = require("path");
-var webpack = require("webpack");
+const path = require('path');
+const webpack = require('webpack');
 
 module.exports = {
   context: __dirname,
-  entry: "./frontend/twinkieandkaren.jsx",
+  entry: './frontend/twinkieandkaren.jsx',
   output: {
     path: path.join(__dirname, 'app', 'assets', 'javascripts'),
-    filename: "bundle.js"
+    filename: 'bundle.js',
   },
   resolve: {
-    extensions: ["", ".js", ".jsx"]
+    extensions: ['', '.js', '.jsx'],
   },
-  plugins:[
-    new webpack.DefinePlugin({
-      'process.env':{
-        'NODE_ENV': JSON.stringify('production')
-      }
-    })
-  ],
   module: {
     loaders: [
       {
         test: /\.jsx?$/,
         exclude: /(node_modules|bower_components)/,
-        loader: 'babel',
+        loader: "babel-loader",
         query: {
-          presets: ['react']
-        }
+          presets: ['react'],
+        },
       },
       {
         test: /\.node$/,
-        loader: "node-loader"
-      }
-    ]
-  }
+        loader: 'node-loader',
+      },
+    ],
+  },
 };
